observatorio/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
celery==5.3.4
redis==5.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.20
email-validator==2.0.0.post2
WTForms==3.0.1
alembic==1.12.0
psycopg2-binary==2.9.7
requests==2.31.0
python-dateutil==2.8.2
